feat: Infile is kinda working
But not very well a problem is that the gyroid is faulty as generated
by csgrs. But also prusa-slicer doesn't handle the notion of "walls" and
then also where the hollow is infilled externally pattern. This maybe
a configuration fault on my part.

For now this doesn't work well enough to ship, and I need to experiment
more probably not doing any infill myself.